Class for gnomon stuff
